% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov2.R
\name{cov2}
\alias{cov2}
\title{sample covariance scaled by n}
\usage{
cov2(x, y)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame}

\item{y}{a numeric vector, matrix or data frame, with compatible dimensions to x}
}
\value{
sample covariance scaled by n, rather than n-1
}
\description{
sample covariance scaled by n
}
\examples{
cov2(c(1,2),c(2,3))
}
