% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_st_Isbell.R
\name{part.st.Isbell}
\alias{part.st.Isbell}
\title{Isbell's spatiotemporal partition}
\usage{
part.st.Isbell(MM, YY, PP = array(1/dim(MM)[1], dim = dim(MM)))
}
\arguments{
\item{MM}{An array contains the monoculture yields, organized by a 3-dimension array: 1-species, 2-time, 3-space.}

\item{YY}{An array contains the mixture yields organized by a 3-dimension array: 1-species, 2-time, 3-space with compatible dimensions to MM.}

\item{PP}{An vector contains the expected relative mixture yields with compatible dimensions to MM. The defalut is PP=array(1/dim(MM)[1],dim=dim(MM)).}
}
\value{
net biodiversity effect, total complementary, non-random overyielding effect, average selection, temporal selection, spatial selection, spatiotemporal selection.
}
\description{
Isbell's spatiotemporal partition
}
\examples{
{
MM = array(c(100,100,100,100,50,50,50,50),dim=c(2,2,2))
YY = array(c(50,50,50,50,50,50,50,50),dim=c(2,2,2))
part.st.Isbell(MM,YY)
}

}
