% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_x_Isbell.R
\name{part.x.Isbell}
\alias{part.x.Isbell}
\title{Isbell's spatiotemporal partition}
\usage{
part.x.Isbell(MM, YY, PP = array(1/dim(MM)[1], dim = dim(MM)), spatial = F)
}
\arguments{
\item{MM}{An array contains the monoculture yields, organized by a 3-dimension array: 1-species, 2-time/space.}

\item{YY}{An array contains the mixture yields organized by a 3-dimension array: 1-species, 2-time/space with compatible dimensions to MM.}

\item{PP}{An vector contains the expected relative mixture yields with compatible dimensions to MM. The defalut is PP=array(1/dim(MM)[1],dim=dim(MM)).}

\item{spatial}{An indecator of spatial/temporal partition. The defalut is spatial=F, i.e. temporal partition.}
}
\value{
net biodiversity effect, total complementary, non-random overyielding effect, average selection, temporal/spatial selection.
}
\description{
Isbell's spatiotemporal partition
}
\examples{
{
MM = array(c(100,100,100,100,50,50,50,50),dim=c(2,4))
YY = array(c(50,50,50,50,50,50,50,50),dim=c(2,4))
part.x.Isbell(MM,YY)
}

}
